Make card component hierarchy similar to GA
