+++ src/app.module.ts
@@ -2,7 +2,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { LineChartModule } from './components/line-chart/line-chart.module';
 
 @NgModule({
   declarations: [
@@ -11,7 +10,6 @@ import { LineChartModule } from './components/line-chart/line-chart.module';
   imports: [
     BrowserModule,
     AppRoutingModule,
-    LineChartModule,
   ],
   bootstrap: [
     AppComponent,
