--- src/components/card-fold/card-fold.component.html
@@ -1,2 +1,0 @@
-<ng-content></ng-content>
-<ng-template></ng-template>
