--- src/components/card-fold/card-fold.component.scss
@@ -1,9 +1,0 @@
-@import "../../utils/constants";
-
-:host {
-  padding: 1.5rem;
-
-  app-card-fold + & {
-    border-top: 1px solid $color-border;
-  }
-}
