--- src/components/card-fold/card-fold.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CardFoldComponent } from './card-fold.component';
-
-@NgModule({
-  declarations: [
-    CardFoldComponent,
-  ],
-  imports: [],
-  exports: [
-    CardFoldComponent,
-  ],
-})
-export class CardFoldModule {
-}
