+++ src/components/card/card.component.scss
@@ -17,5 +17,24 @@
     align-items: stretch;
     border-radius: 2px;
     background-color: white;
+
+    .tabs {
+      display: flex;
+      align-items: stretch;
+      height: 4rem;
+
+      .tab {
+        flex: 1;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        cursor: pointer;
+
+        &.active {
+          border-top: 4px solid $color-highlight;
+          font-weight: 500;
+        }
+      }
+    }
   }
 }
