+++ src/components/card/card.module.ts
@@ -1,11 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CardComponent } from './card.component';
+import { LineChartModule } from '../line-chart/line-chart.module';
+import { AudificationModule } from '../../directives/audification/audification.module';
+import { CommonModule } from '@angular/common';
 
 @NgModule({
   declarations: [
     CardComponent,
   ],
-  imports: [],
+  imports: [
+    LineChartModule,
+    AudificationModule,
+    CommonModule,
+  ],
   exports: [
     CardComponent,
   ],
