+++ src/components/dashboard/dashboard.component.html
@@ -39,7 +39,7 @@
 </aside>
 <div class="card-container">
   <div class="row">
-    <app-line-chart appAudification></app-line-chart>
-    <app-line-chart></app-line-chart>
+    <app-card title="Line Chart" [tabbed]="true" type="line"
+              [measureNames]="['activeUsers', 'revenue', 'eventCount']"></app-card>
   </div>
 </div>
