+++ src/components/dashboard/dashboard.component.scss
@@ -54,6 +54,7 @@
     overflow-y: auto;
 
     .row {
+      flex-shrink: 0;
       display: flex;
       justify-content: center;
       flex-wrap: wrap;
