+++ src/components/dashboard/dashboard.module.ts
@@ -1,21 +1,19 @@
 import { NgModule } from '@angular/core';
 import { DashboardComponent } from './dashboard.component';
-import { LineChartModule } from '../line-chart/line-chart.module';
 import { PreferenceModule } from '../preference/preference.module';
 import { FormsModule } from '@angular/forms';
 import { SwitchModule } from '../switch/switch.module';
-import { AudificationModule } from '../../directives/audification/audification.module';
+import { CardModule } from '../card/card.module';
 
 @NgModule({
   declarations: [
     DashboardComponent,
   ],
   imports: [
-    LineChartModule,
     PreferenceModule,
     FormsModule,
     SwitchModule,
-    AudificationModule,
+    CardModule,
   ],
   exports: [
     DashboardComponent,
