--- src/components/line-chart-visualization/line-chart-visualization.component.html
@@ -1,3 +1,0 @@
-<div class="active-indicator">
-  {{formattedActiveDatum}}
-</div>
