--- src/components/line-chart-visualization/line-chart-visualization.component.scss
@@ -1,17 +1,0 @@
-@import "../../utils/constants";
-
-:host {
-  position: relative;
-  display: block;
-
-  .active-indicator {
-    @include card-shadow;
-    position: absolute;
-    font-weight: 500;
-    padding: 1rem;
-    background-color: white;
-    white-space: nowrap;
-    color: $color-highlight;
-    opacity: 0;
-  }
-}
