+++ src/components/line-chart/line-chart.component.scss
@@ -1,24 +1,18 @@
 @import "../../utils/constants";
 
 :host {
-  .card {
-    .measures {
-      display: flex;
-      align-items: stretch;
-      height: 4rem;
+  position: relative;
+  display: block;
+  padding: 1.5rem;
 
-      .measure {
-        flex: 1;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        cursor: pointer;
-
-        &.active {
-          border-top: 4px solid $color-highlight;
-          font-weight: 500;
-        }
-      }
-    }
+  .active-indicator {
+    @include card-shadow;
+    position: absolute;
+    font-weight: 500;
+    padding: 1rem;
+    background-color: white;
+    white-space: nowrap;
+    color: $color-highlight;
+    opacity: 0;
   }
 }
