+++ src/components/line-chart/line-chart.module.ts
@@ -1,21 +1,17 @@
 import { NgModule } from '@angular/core';
 import { LineChartComponent } from './line-chart.component';
-import { LineChartVisualizationModule } from '../line-chart-visualization/line-chart-visualization.module';
-import { CardModule } from '../card/card.module';
+import { DataModule } from '../../services/data/data.module';
 import { CommonModule } from '@angular/common';
 import { A11yHostModule } from '../../directives/a11y-host/a11y-host.module';
-import { CardFoldModule } from '../card-fold/card-fold.module';
 
 @NgModule({
   declarations: [
     LineChartComponent,
   ],
   imports: [
-    CardModule,
-    LineChartVisualizationModule,
-    A11yHostModule,
-    CardFoldModule,
+    DataModule,
     CommonModule,
+    A11yHostModule,
   ],
   exports: [
     LineChartComponent,
