+++ src/d3/base.d3.ts
@@ -15,8 +15,6 @@ export interface RenderOptions {
 
 export abstract class BaseD3<T extends RenderOptions> {
   protected colorHighlight = 'rgb(66, 133, 244)';
-
-  protected svg: d3.Selection<SVGSVGElement, unknown, null, undefined>;
   private clear$?: Subject<undefined>;
 
   constructor(protected renderOptions: T) {
@@ -30,6 +28,10 @@ export abstract class BaseD3<T extends RenderOptions> {
     return d3.select(this.renderOptions.elementRef.nativeElement);
   }
 
+  protected get svg() {
+    return this.container.select('svg');
+  }
+
   config(renderOptions: T) {
     this.renderOptions = renderOptions;
     return this;
@@ -41,8 +43,7 @@ export abstract class BaseD3<T extends RenderOptions> {
     this.clear();
     this.clear$ = new Subject();
 
-    this.svg = this.container
-      .append('svg')
+    this.svg
       .style('width', width)
       .style('height', height)
       .attr('viewBox', [0, 0, width, height].join(' '));
